Moved interpret functions to utils
